[RHEL5] testing on RHEL5 box

+++ scripts/prepare_project.sh
@@ -5,10 +5,12 @@
 langs=$1
 stamp=$2
 
-if [ -e publican.cfg ]; then
-    sed -e "s/brand:.*//" publican.cfg > publican.cfg.striped
-    publican update_pot --config publican.cfg.striped \
-    && publican update_po --config publican.cfg.striped --langs ${langs}
+if [ -e /usr/bin/publican ] ; then
+    if [ -e publican.cfg ]; then
+	sed -e "s/brand:.*//" publican.cfg > publican.cfg.striped
+	publican update_pot --config publican.cfg.striped \
+	&& publican update_po --config publican.cfg.striped --langs ${langs}
+    fi
 fi
 
 touch ${stamp}
